+++ Release_Notes.md
@@ -1,5 +1,10 @@
 # Aio Database Release Notes
 
+### v5.3
+- **insert_value** is now accepting references
+- Add **Contains**, **StartsWith** and **EndsWith** Operators
+- Internal: Added test to guarantee that all apis work before release
+
 ### v5.2 
 - Fix Documentation
 - Fix Some Examples showing the same code for in-memory and local db